pyproject.toml
[tool.poetry]
name = "nexus_test"
version = "0.1.0"
description = ""
authors = ["Kevin Leptons <kevin.leptons@gmail.com>"]

[tool.poetry.dependencies]
python = "3.7.*"
tesserocr = "^2.5.1"
click = "^7.1.2"
pdf2image = "^1.13.1"
filetype = "^1.0.7"
opencv-python = "4.2.0.34"
imutils = "^0.5.3"
pyenchant = "^3.1.1"

[tool.poetry.dev-dependencies]
yapf = "^0.30.0"
isort = "^5.3.2"
